% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parameter_summaries.r
\name{add_parameter_summaries}
\alias{add_parameter_summaries}
\title{Adds parameter summary slots to a MoBAMAResult object}
\usage{
add_parameter_summaries(result)
}
\arguments{
\item{result}{The MoBAMAResult object to fill with parameter summaries}
}
\value{
the filled MoBAMAResult object
}
\description{
Adds parameter summary slots to a MoBAMAResult object
}
