% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen_barplot.r
\name{antigen_barplot}
\alias{antigen_barplot}
\title{Plot the antigen means as a barplot}
\usage{
antigen_barplot(result, includeIntervals = F)
}
\arguments{
\item{result}{The MoBAMAResult object.}

\item{includeIntervals}{A boolean indicating whether to
include sampling intervals for the antigen means.
Defaults to \code{FALSE}.}
}
\value{
A ggplot barplot object.
}
\description{
This function plots the antigen means from a model fit as a
  barplot.
}
