% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen_barplot_custom.r
\name{antigen_barplot_custom}
\alias{antigen_barplot_custom}
\title{Plot the antigen means as a barplot}
\usage{
antigen_barplot_custom(result, agOrderTable, agTitle = "Antigen",
  meanTitle = "Mean", colorTitle = "Timepoint", aglines = "white",
  incIntervals = F, intervalSize = 3, intervalBarSize = 1.5,
  intervalColor = "gray30")
}
\arguments{
\item{result}{The MoBAMAResult object.}

\item{includeIntervals}{A boolean indicating whether to
include sampling intervals for the antigen means.
Defaults to \code{FALSE}.}
}
\value{
A ggplot barplot object.
}
\description{
This function plots the antigen means from a model fit as a
  barplot.
TODO: add more options
      tp filtering
      colors (by antigen group, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?
}
