% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.r
\name{extractParams}
\alias{extractParams}
\title{Extracts parameters from the model results}
\usage{
extractParams(object, varStr, fullname = FALSE)
}
\arguments{
\item{object}{the MoBAMAResult object to query}

\item{varStr}{the string which must match either the start or all of the parameter name}

\item{fullname}{indicates whether the whole variable name must be matched, or just the start}
}
\value{
a \link{data.frame} containing the extracted parameter information
}
\description{
Extracts parameters from the model results that match
the string provided
}
