% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.r
\name{extractParams2}
\alias{extractParams2}
\title{Extracts parameters from the model results}
\usage{
extractParams2(object, startStr, endStr)
}
\arguments{
\item{object}{the MoBAMAResult object to query}

\item{startStr}{the string which must match the start of the parameter name}

\item{endStr}{the string which must match the end of the parameter name}
}
\value{
a \link{data.frame} containing the extracted parameter information
}
\description{
Extracts parameters from the model results that match
the start and ending strings provided
}
