% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBAMA-package.R
\docType{package}
\name{MoBAMA-package}
\alias{MoBAMA-package}
\alias{MoBAMA}
\title{BAMA Data Modeling via RStan}
\description{
The \pkg{MoBAMA} package is used to model BAMA and Fc array data using a Bayesian
grouped mixture model.
}
