% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBAMA.r
\name{MoBAMA}
\alias{MoBAMA}
\title{Runs a model on BAMA or Fc array data using stan}
\usage{
MoBAMA(data, dataType = "fc", nIter = 2000, nChains = 1,
  outFolder = NULL, outFile = date_filename("MoBAMA_stanfit.rds"), ...)
}
\arguments{
\item{dataType}{type of data to be modeled ('bama' or 'fc')}

\item{nIter}{the number of iterations per chain for the stan sampling}

\item{nChains}{the number of chains to run for the stan sampling}

\item{outFolder}{the folder to save the stan results to (or NULL to not save)}

\item{outFile}{the filename to save the stan results to -
defaults to '<yyyy-mm-dd>_MoBAMA_stanfit.rds'}

\item{...}{additional parameters to pass to the stan function}

\item{bamaData}{the data to be modeled}
}
\value{
a summary of the model results
}
\description{
Runs a chosen grouped mixture model on the provided data.
TODO: change params out to just list all the options?
      saving the stanfit object or returning it
      returning the summary in a better format
}
