% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_barplot_custom.r
\name{parameter_barplot_custom}
\alias{parameter_barplot_custom}
\title{Plot the parameter data as a barplot}
\usage{
parameter_barplot_custom(paramData, paramLabel = "Parameter",
  valueLabel = "Value", fillLabel = "Fill", groupSepColor = NULL,
  incIntervals = F, intervalPtSize = 3, intervalBarSize = 1.5,
  intervalColor = "gray30")
}
\arguments{
\item{paramData}{A \code{data.frame} with all the data necessary for
creating the barplot. This includes the following columns:
\describe{
\item{order}{The order of parameters on the axis. Should include all
integers from 1 to the number of rows in the dataset.}
\item{label}{The parameter label to use in the barplot. Consecutive
parameters with the same label will share a label on the axis rather
than repeating the label.}
\item{labelColor}{The color to use for the label text.}
\item{paramValue}{The value of the parameter (length of the bar).}
\item{qlower}{The lower sampling percentile for the value of the
parameter. Required if \code{incIntervals = TRUE}.}
\item{qupper}{The upper sampling percentile for the value of the
parameter. Required if \code{incIntervals = TRUE}.}
\item{group}{The parameter group. Each distinct group should cover
a contiguous set of parameters by order in order to separate groups
by lines. Required if \code{groupSepColor != NULL}.}
\item{fillGroup}{The group designation for bar fill color.}
\item{fillColor}{The fill color for the group.}
}}

\item{paramLabel}{The label for the parameter axis. Defaults to 'Parameter'.}

\item{valueLabel}{The label for the value axis. Defaults to 'Value'.}

\item{fillLabel}{The label for the fill color legend. Defaults to 'Fill'.}

\item{groupSepColor}{A string defining the color for lines separating groups
on the parameter axis or \code{NULL} for no lines. Defaults to \code{NULL}.}

\item{incIntervals}{A boolean indicating whether to include sampling
intervals for the parameter values. Defaults to \code{FALSE}.}

\item{intervalPtSize}{The size for the center point of the
sampling interval. Defaults to 3.}

\item{intervalBarSize}{The size for the sampling interval bars.
Defaults to 1.5}

\item{intervalColor}{A string defining the color for the sampling interval
error bars. Defaults to 'gray30'.}
}
\value{
A ggplot barplot object.
}
\description{
This function plots sampled parameter values from a model fit as a
barplot. All sorting and filtering is done in the passed data.
}
