% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBAMAResult-methods.r
\name{print.MoBAMAResult}
\alias{print.MoBAMAResult}
\title{Print a MoBAMAResult Object}
\usage{
\method{print}{MoBAMAResult}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MoBAMAResult}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
This function prints basic information about the model fit by a
\code{\link{MoBAMA}} call.
}
\examples{
print(MB)
}
