% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reagent_barplot.r
\name{reagent_barplot}
\alias{reagent_barplot}
\title{Plot the reagent means as a barplot}
\usage{
reagent_barplot(result, includeIntervals = F)
}
\arguments{
\item{result}{The MoBAMAResult object.}

\item{includeIntervals}{A boolean indicating whether to
include sampling intervals for the reagent means.
Defaults to \code{FALSE}.}
}
\value{
A ggplot barplot object.
}
\description{
This function plots the reagent means from a model fit as a
  barplot.
}
