% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reagent_barplot_custom.r
\name{reagent_barplot_custom}
\alias{reagent_barplot_custom}
\title{Plot the reagent means as a barplot}
\usage{
reagent_barplot_custom(result, reOrderTable, reText = "Fc Variable",
  meanText = "Mean", colorText = "Timepoint", relines = "white",
  incIntervals = F, intervalPtSize = 3, intervalBarSize = 1.5,
  intervalColor = "gray30")
}
\arguments{
\item{result}{The MoBAMAResult object.}

\item{reOrderTable}{A \code{data.frame} with all re/tp combinations
to include as well as ordering labeling and color information.
Should have the following columns:
re, tp, order, label, group, color, barlabel, barcolor}

\item{reText}{The label for the Fc axis. Defaults to 'Fc Variable'.}

\item{meanText}{The label for the Fc mean axis. Defaults to 'Mean'.}

\item{colorText}{The label for the color legend. Defaults to 'Timepoint'.}

\item{incIntervals}{A boolean indicating whether to include sampling
intervals for the Fc variable means. Defaults to \code{FALSE}.}

\item{intervalPtSize}{The size for the center point of the
sampling interval. Defaults to 3.}

\item{intervalBarSize}{The size for the sampling interval bars.
Defaults to 1.5}

\item{intervalColor}{A string defining the color for the sampling interval
error bars. Defaults to 'gray30'.}

\item{reLines}{A string defining the color for lines separating groups
(by label) on the Fc axis or \code{NULL} for no lines.
Defaults to 'white'.}
}
\value{
A ggplot barplot object.
}
\description{
This function plots the reagent means from a model fit as a
barplot with additional options for sorting and filtering the
data displayed.
}
