% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_call_evaluation.r
\name{response_call_evaluation}
\alias{response_call_evaluation}
\title{Summarize the responses from analyzing (particularly) BAMA data
into a table of confusion matrices}
\usage{
response_call_evaluation(result, subjectData, fdrs, cutoffs)
}
\arguments{
\item{result}{The MoBAMAResult object.}

\item{subjectData}{A \code{data.frame} with columns subjectId and isResponder
defining which subjects are believed to be responders.}

\item{fdrs}{A vector of target false-discovery rates.}

\item{cutoffs}{A vector of cutoff probabilities for calling responses.}
}
\value{
A \code{data.frame} of confusion matrix entries with columns
  ag, timepoint, fdr, cutoffProb, tp, fp, tn, fn
}
\description{
This function calculates confusion matrices from a response dataset and truth
based on subjects (each subject considered a responder or not) for a series
of cutoff probabilities for calling responses and false-discovery rates.
Subsets on individual antigen and timepoints as well as aggregates.
}
