% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_heatmap.r
\name{response_heatmap}
\alias{response_heatmap}
\title{Plots the response probabilities for all observations as a heatmap
TODO: add more options
        tp filtering?
        sorting axes
        colors (by group, ag/re?, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?}
\usage{
response_heatmap(result, responseThreshold = NULL, groupby = "ag")
}
\arguments{
\item{result}{the MoBAMAResult object}

\item{responseThreshold}{if not NULL, the threshold probability defining a response,
resulting in a two-color heatmap rather than a continuous heatmap}

\item{groupby}{the field (ag or re) to group by}
}
\value{
a ggplot plot object
}
\description{
Plots the response probabilities for all observations as a heatmap
TODO: add more options
        tp filtering?
        sorting axes
        colors (by group, ag/re?, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?
}
