% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.r
\name{responses}
\alias{responses}
\title{Get the inferred response probabilities for all observations from
a model fit}
\usage{
responses(result)
}
\arguments{
\item{result}{The MoBAMAResult object to query.}
}
\value{
A \code{data.frame} containing the observation and response
  probability information.
}
\description{
This function adds response probabilities from a model fit
to the modeled dataset.
}
