% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBAMAResult-methods.r
\name{summary.MoBAMAResult}
\alias{summary.MoBAMAResult}
\title{Summarize a MoBAMAResult Object}
\usage{
summary.MoBAMAResult(object, ...)
}
\arguments{
\item{object}{An object of class \code{MoBAMAResult}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
This function prints basic information about the model fit by a
\code{\link{MoBAMA}} call.
}
\examples{
print(MB)
}
