% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen_barplot.r
\name{antigen_barplot}
\alias{antigen_barplot}
\title{Barplot of antigen means
TODO: add more options
        colors (by antigen group, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?}
\usage{
antigen_barplot(result, includeIntervals = F)
}
\arguments{
\item{result}{the MoBAMAResult object}

\item{includeIntervals}{include sampling intervals for the probabilities.
Defaults to false because of identifiability issues}
}
\value{
a ggplot plot object
}
\description{
Barplot of antigen means
TODO: add more options
        colors (by antigen group, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?
}
