% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_cutoff.r
\name{fdr_cutoff}
\alias{fdr_cutoff}
\title{Calculate the cutoff probability for a requested false-discovery rate}
\usage{
fdr_cutoff(responseProbs, fdr = 0.05)
}
\arguments{
\item{responseProbs}{The response probabilities to use for calculating
the minimum cutoff probability.}

\item{fdr}{The requested false-discovery rate.}
}
\value{
The maximum probability cutoff for defining a response that
  will result in an expected false-discovery rate equal to or less
  than that requested.
}
\description{
This function calculates the minimum cutoff probability for defining a
response from a set of response probabilities such that the expected
false-discovery rate is no greater than the requested FDR.
}
