% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBAMA-package.R
\docType{package}
\name{MoBAMA-package}
\alias{MoBAMA-package}
\title{MoBAMA (Modeling of Binding Antibody Multiplex Assays)}
\description{
This package fits a Bayesian grouped mixture model for BAMA and
Fc array data using stan. The model identifies responses in the
data.
}
\seealso{
\itemize{
    \item \code{\link{MoBAMA}}, for the main model fitting routine.
  }
}
