% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reagent_barplot.r
\name{reagent_barplot}
\alias{reagent_barplot}
\title{Barplot of Fc variable means
TODO: add more options
        colors (by Fc group, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?}
\usage{
reagent_barplot(result, includeIntervals = F)
}
\arguments{
\item{result}{the MoBAMAResult object}

\item{includeIntervals}{include sampling intervals for the probabilities.
Defaults to false because of identifiability issues}
}
\value{
a ggplot plot object
}
\description{
Barplot of Fc variable means
TODO: add more options
        colors (by Fc group, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?
}
