% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_heatmap.r
\name{response_heatmap}
\alias{response_heatmap}
\title{Plot the response probabilities for all observations as a heatmap}
\usage{
response_heatmap(result, responseThreshold = NULL, groupby = "ag")
}
\arguments{
\item{result}{The MoBAMAResult object.}

\item{responseThreshold}{If not NULL, the threshold probability
defining a response, resulting in a two-color heatmap rather than
a continuous heatmap. Defaults to \code{NULL}.}

\item{groupby}{The field ('ag' or 're') to group observations by.
Defaults to 'ag'.}
}
\value{
A ggplot heatmap.
}
\description{
This function plots the response probabilities from a model fit as a
  heatmap.
TODO: add more options
        tp filtering?
        sorting axes
        colors (by group, ag/re?, user-specified)
      make the plot cleaner/nicer
      not hard code axis labels?
}
