% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_heatmap_custom.r
\name{response_heatmap_custom}
\alias{response_heatmap_custom}
\title{Plots the response probabilities for all observations as a heatmap
with additional options for sorting and filtering both axes
TODO: implement color}
\usage{
response_heatmap_custom(result, xorderTable, yorderTable,
  responseThreshold = NULL, xtext = "Antigen/Fc Variable",
  xlines = "white", ytext = "SubjectId", ylines = NULL)
}
\arguments{
\item{result}{the MoBAMAResult object}

\item{xorderTable}{a table with all ag/re/tp combinations to include
as well as ordering, labeling, and color information.
Should have the following columns: ag, re, tp, order, label, color}

\item{yorderTable}{a table with all subjectIds to include
as well as ordering, labeling, and color information.
Should have the following columns: subjectId, order, label, color}

\item{responseThreshold}{if not NULL, the threshold probability defining a response,
resulting in a two-color heatmap rather than a continuous heatmap}

\item{xtext}{the label for the x-axis}

\item{xlines}{the color for lines separating groups (by label) on the x-axis
or NULL for no lines}

\item{ytext}{the label for the y-axis}

\item{ylines}{the color for lines separating groups (by label) on the y-axis
or NULL for no lines}
}
\value{
a ggplot plot object
}
\description{
Plots the response probabilities for all observations as a heatmap
with additional options for sorting and filtering both axes
TODO: implement color
}
