% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.r
\name{responses}
\alias{responses}
\title{Gets the inferred response probabilities for observations from
a MoBAMAResult object}
\usage{
responses(result)
}
\arguments{
\item{result}{the MoBAMAResult object to query}
}
\value{
a \code{data.frame} containing the response probability information
}
\description{
Gets the inferred response probabilities for observations from
a MoBAMAResult object
}
